backend/users-ms/src/main/java/com/azat4dev/demobooking/users/users_commands/data/repositories/DomainEventSerializerImpl.java
package com.azat4dev.demobooking.users.users_commands.data.repositories;

import com.azat4dev.demobooking.common.domain.event.DomainEventNew;
import com.azat4dev.demobooking.users.users_commands.data.repositories.dto.*;
import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.JsonToken;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.SerializationFeature;
import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
import com.fasterxml.jackson.databind.module.SimpleModule;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;

import java.io.IOException;
import java.time.LocalDateTime;
import java.util.HashMap;
import java.util.Map;


public class DomainEventSerializerImpl implements DomainEventSerializer {

    private final ObjectMapper objectMapper;

    public DomainEventSerializerImpl() {

        final var objectMapper = new ObjectMapper();

        objectMapper.registerModule(new JavaTimeModule());

        this.objectMapper = objectMapper.configure(SerializationFeature.WRITE_DATE_TIMESTAMPS_AS_NANOSECONDS, true)
            .configure(DeserializationFeature.READ_DATE_TIMESTAMPS_AS_NANOSECONDS, true);
    }

    @Override
    public String serialize(DomainEventNew<?> event) {
        try {
            return objectMapper.writeValueAsString(DomainEventDTO.makeFrom(event));
        } catch (IOException e) {
            throw new RuntimeException(e);
        }
    }

    @Override
    public DomainEventNew<?> deserialize(String event) {
        try {
            final var dto = this.objectMapper.readValue(event, DomainEventDTO.class);
            return dto.toDomain();
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}

